Move audit log view to own file.
